[POT-67][Refactor] 지도 기능 모듈화
